requirements.txt
clamd==1.0.2
Flask==2.1.0
google-cloud-storage==1.37.1
